+++ src/ladb_opencutlist/ruby/constants.rb
@@ -1,7 +1,7 @@
 module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
-    EXTENSION_VERSION = '1.8.4-dev'.freeze
-    EXTENSION_BUILD = '201912191201'.freeze
+    EXTENSION_VERSION = '1.8.4'.freeze
+    EXTENSION_BUILD = '202001091041'.freeze
 
 end