+++ src/ladb_opencutlist/ruby/tool/highlight_part_tool.rb
@@ -87,6 +87,10 @@ module Ladb::OpenCutList
       view.invalidate
     end
 
+    def suspend(view)
+      view.invalidate
+    end
+
     def resume(view)
       view.invalidate
     end